physics test cleanup
