+++ tests/physics/src/ShapeColliderTests.cpp
@@ -10,7 +10,6 @@
 //
 
 //#include <stdio.h>
-#include "PhysicsTestUtil.h"
 #include <iostream>
 #include <math.h>
 
@@ -29,10 +28,10 @@
 #include "ShapeColliderTests.h"
 
 
-//const glm::vec3 origin(0.0f);
-//static const glm::vec3 xAxis(1.0f, 0.0f, 0.0f);
-//static const glm::vec3 yAxis(0.0f, 1.0f, 0.0f);
-//static const glm::vec3 zAxis(0.0f, 0.0f, 1.0f);
+const glm::vec3 origin(0.0f);
+static const glm::vec3 xAxis(1.0f, 0.0f, 0.0f);
+static const glm::vec3 yAxis(0.0f, 1.0f, 0.0f);
+static const glm::vec3 zAxis(0.0f, 0.0f, 1.0f);
 
 QTEST_MAIN(ShapeColliderTests)
